user_management/requirements.txt
flask
gunicorn
pyJWT
gevent
Flask-Cors
psycopg2
psycogreen
python-dotenv
Flask-pydantic
sqlalchemy
cryptography
